requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
pyyaml>=6.0.1
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
torch>=2.1.0
torchvision>=0.16.0
onnxruntime>=1.16.0
